Move all public enums to a separate header
